+++ src/ScreenCalibrationWidget.h
@@ -41,7 +41,7 @@ protected:
 
 	QDoubleSpinBox * _sbDPI;
 	QRect _rulerRect;
-	unsigned int _majorTickHeight{20}, _mediumTickHeight{10}, _minorTickHeight{5}, _paperTickHeight{40};
+	int _majorTickHeight{20}, _mediumTickHeight{10}, _minorTickHeight{5}, _paperTickHeight{40};
 	QMenu _contextMenu;
 	QActionGroup _contextMenuActionGroup;
 	QSignalMapper _unitSignalMapper;
@@ -60,11 +60,11 @@ protected:
 		bool visible;
 	};
 	QList<paperSize> _paperSizes;
-	unsigned int _curUnit;
+	int _curUnit;
 	int _hSpace{0};
 
 	QPoint _mouseDownPos;
-	float _mouseDownInches{0};
+	double _mouseDownInches{0};
 	bool _isDragging{false};
 };
